♻️ : cartUtils 순수함수 분리
