+++ src/refactoring/hooks/utils/cartUtils.ts
@@ -1,67 +1,54 @@
 import { CartItem, Coupon } from "../../../types";
+import { getCouponDiscount } from "./couponUtils";
+import { getMaxApplicableDiscount, getTotalAfterProductDiscount, getTotalBeforeProductDiscount } from "./discountUtils";
 
 export const calculateItemTotal = (item: CartItem) => {
-    const basePrice = item.product.price * item.quantity;
-    const discountRate = 1 - getMaxApplicableDiscount(item);
+  const basePrice = item.product.price * item.quantity;
+  const discountRate = 1 - getMaxApplicableDiscount(item);
+
+  return basePrice * discountRate;
+};
   
-    return basePrice * discountRate;
+export const calculateCartTotal = (
+  cart: CartItem[],
+  selectedCoupon: Coupon | null
+) => {
+  const totalBeforeDiscount = getTotalBeforeProductDiscount(cart)
+  const totalAfterProductDiscount = getTotalAfterProductDiscount(cart);
+
+  const couponDiscount = getCouponDiscount(totalAfterProductDiscount, selectedCoupon);
+
+  const totalAfterDiscount = totalAfterProductDiscount - couponDiscount;
+  const totalDiscount = totalBeforeDiscount - totalAfterDiscount;
+
+  return {
+    totalBeforeDiscount,
+    totalAfterDiscount,
+    totalDiscount,
   };
-  
-  export const getMaxApplicableDiscount = (item: CartItem) => {
-    return item.product.discounts.reduce((maxRate, discount) => {
-      if (item.quantity >= discount.quantity) {
-        return Math.max(maxRate, discount.rate);
-      }
-  
-      return maxRate;
-    }, 0);
-  };
-  
-  export const calculateCartTotal = (
-    cart: CartItem[],
-    selectedCoupon: Coupon | null
-  ) => {
-    const totalBeforeDiscount = cart.reduce((acc, item) => {
-      return acc + item.product.price * item.quantity;
-    }, 0);
-  
-    const totalAfterProductDiscount = cart.reduce((acc, item) => {
-      return acc + calculateItemTotal(item);
-    }, 0);
-  
-    const couponDiscount = () => {
-      if (!selectedCoupon) return 0;
-  
-      if (selectedCoupon.discountType === "percentage") {
-        return totalAfterProductDiscount * (selectedCoupon.discountValue / 100);
-      }
-  
-      return selectedCoupon.discountValue;
-    }
-  
-    const totalAfterDiscount = totalAfterProductDiscount - couponDiscount();
-    const totalDiscount = totalBeforeDiscount - totalAfterDiscount;
-  
-    return {
-      totalBeforeDiscount,
-      totalAfterDiscount,
-      totalDiscount,
-    };
-  };
-  
-  export const updateCartItemQuantity = (
-    cart: CartItem[],
-    productId: string,
-    newQuantity: number
-  ): CartItem[] => {
-    if (newQuantity === 0) {
-      return cart.filter((item) => item.product.id !== productId);
-    }
-  
-    return cart.map((item) => {
-      if (item.product.id !== productId) return item;
-  
-      const limitedQuantity = Math.min(newQuantity, item.product.stock);
-      return { ...item, quantity: limitedQuantity };
-    });
-  };
+};
+
+const getCartRemoveItem = (cart: CartItem[], productId: string) => {
+  return cart.filter((item) => item.product.id !== productId);
+}
+
+const getCartLimitedQuantity = (cart: CartItem[], productId: string, newQuantity: number) => {
+  return cart.map((item) => {
+    if (item.product.id !== productId) return item;
+
+    const limitedQuantity = Math.min(newQuantity, item.product.stock);
+    return { ...item, quantity: limitedQuantity };
+  })
+}
+
+export const updateCartItemQuantity = (
+  cart: CartItem[],
+  productId: string,
+  newQuantity: number
+): CartItem[] => {
+  if (newQuantity === 0) {
+    return getCartRemoveItem(cart, productId);
+  }
+
+  return getCartLimitedQuantity(cart, productId, newQuantity);
+};
