+++ src/refactoring/hooks/utils/couponUtils.ts
@@ -1,0 +1,11 @@
+import { Coupon } from "../../../types";
+
+export const getCouponDiscount = (totalAfterProductDiscount: number, selectedCoupon: Coupon | null) => {
+  if (!selectedCoupon) return 0;
+
+  if (selectedCoupon.discountType === "percentage") {
+      return totalAfterProductDiscount * (selectedCoupon.discountValue / 100);
+  }
+
+  return selectedCoupon.discountValue;
+}
